chore: Reorganize API implementations
