+++ CHANGELOG.md
@@ -7,10 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Changed
-- The `DB` environment variable is now ignored. All persistent data now lives on the database identified by the value of the `DATABASE_URL` environment variable. [See the README](/server/README.md) for details.
+- The server's `DB` environment variable is now ignored. All persistent data now lives on the database identified by the value of the server's `DATABASE_URL` environment variable. [See the README](/server/README.md) for details.
+- Reorganized API implementations for better parity between [SvelteKit](https://kit.svelte.dev) and Express server.
 
 ### Removed
-- The database migration logic from [0.13.0](#0130---2022-10-01) has been removed. Your old data should now live on a MySQL server of your choosing!
+- The database migration logic from [0.13.0](#0130---2022-10-01) has been removed. Your old data now lives on a MySQL server of your choosing.
 
 ## [0.13.1] - 2022-10-01
 ### Fixed
