+++ server/routes/v0/+server.ts
@@ -1,0 +1,3 @@
+import type { RequestHandler } from "express";
+
+export const GET: RequestHandler = (req, res) => res.json({ message: "lol" });
