+++ server/routes/v0/db/users/[uid]/[collectionId]/+server.ts
@@ -1,11 +1,12 @@
 import type { Params } from "./Params.js";
 import type { Request, Response } from "express";
 import { NotFoundError } from "../../../../../../errors/index.js";
+import { respondData, respondSuccess } from "../../../../../../responses.js";
 import { statsForUser } from "../../../../../../database/io.js";
-import { respondSuccess } from "../../../../../../responses.js";
 import {
 	CollectionReference,
 	deleteCollection,
+	getCollection,
 	isCollectionId,
 } from "../../../../../../database/index.js";
 
@@ -17,7 +18,15 @@ function collectionRef(req: Request<Params>): CollectionReference | null {
 	return new CollectionReference(uid, collectionId);
 }
 
-export async function deleteDataCollection(req: Request<Params>, res: Response): Promise<void> {
+export async function GET(req: Request<Params>, res: Response): Promise<void> {
+	const ref = collectionRef(req);
+	if (!ref) throw new NotFoundError();
+
+	const items = await getCollection(ref);
+	respondData(res, items);
+}
+
+export async function DELETE(req: Request<Params>, res: Response): Promise<void> {
 	const uid = req.params.uid;
 
 	const ref = collectionRef(req);
