+++ server/routes/v0/db/users/[uid]/attachments/[documentId]/blob/[fileName]/+server.ts
@@ -1,11 +1,21 @@
-import type { Params } from "./Params.js";
+import type { DocumentData } from "../../../../../../../../../database/schemas.js";
+import type { Params } from "./Params";
 import type { Request, RequestHandler, Response } from "express";
-import { BadRequestError, NotEnoughRoomError } from "../../../../../../../../../errors/index.js";
 import { maxSpacePerUser, MAX_FILE_BYTES } from "../../../../../../../../../auth/limits.js";
-import { respondSuccess } from "../../../../../../../../../responses.js";
+import { respondData, respondSuccess } from "../../../../../../../../../responses.js";
 import { sep as pathSeparator } from "node:path";
 import { simplifiedByteCount } from "../../../../../../../../../transformers/simplifiedByteCount.js";
-import { statsForUser, upsertFileData } from "../../../../../../../../../database/io.js";
+import {
+	destroyFileData,
+	fetchFileData,
+	statsForUser,
+	upsertFileData,
+} from "../../../../../../../../../database/io.js";
+import {
+	BadRequestError,
+	NotEnoughRoomError,
+	NotFoundError,
+} from "../../../../../../../../../errors/index.js";
 import multer, { memoryStorage } from "multer";
 
 /**
@@ -40,6 +50,41 @@ function requireFilePathParameters(params: Params): Required<Params> {
 	};
 }
 
+export async function DELETE(req: Request<Params>, res: Response): Promise<void> {
+	const userId = req.params.uid;
+
+	const { fileName } = requireFilePathParameters(req.params);
+	await destroyFileData(userId, fileName);
+
+	// Report the user's new usage
+	const { totalSpace, usedSpace } = await statsForUser(userId);
+	const userSizeDesc = simplifiedByteCount(usedSpace);
+	const maxSpacDesc = simplifiedByteCount(maxSpacePerUser);
+	console.debug(`User ${userId} has used ${userSizeDesc} of ${maxSpacDesc}`);
+
+	// When done, get back to the caller with new stats
+	respondSuccess(res, { totalSpace, usedSpace });
+}
+
+interface FileData {
+	contents: string;
+	_id: string;
+}
+
+export async function GET(req: Request<Params>, res: Response): Promise<void> {
+	const { uid: userId, fileName } = requireFilePathParameters(req.params);
+
+	const file = await fetchFileData(userId, fileName);
+	if (file === null) throw new NotFoundError();
+
+	const contents = file.contents.toString("utf8");
+	const fileData: DocumentData<FileData> = {
+		contents,
+		_id: req.params.fileName,
+	};
+	respondData(res, fileData);
+}
+
 export const upload = multer({
 	storage: memoryStorage(),
 	limits: {
@@ -48,7 +93,7 @@ export const upload = multer({
 	},
 }).single("file") as RequestHandler<Params & { [key: string]: string }>;
 
-export async function postFileBlob(req: Request<Params>, res: Response): Promise<void> {
+export async function POST(req: Request<Params>, res: Response): Promise<void> {
 	if (!req.file) throw new BadRequestError("You must include a file to store");
 
 	const userId = req.params.uid;
