--- server/routes/v0/index.ts
@@ -1,3 +1,0 @@
-import type { RequestHandler } from "express";
-
-export const lol: RequestHandler = (req, res) => res.json({ message: "lol" });
