+++ server/routes/v0/logout/+server.ts
@@ -1,8 +1,8 @@
 import type { Request, Response } from "express";
-import { addJwtToBlacklist, jwtTokenFromRequest } from "../../auth/jwt.js";
-import { respondSuccess } from "../../responses.js";
+import { addJwtToBlacklist, jwtTokenFromRequest } from "../../../auth/jwt.js";
+import { respondSuccess } from "../../../responses.js";
 
-export function postLogout(req: Request, res: Response): void {
+export function POST(req: Request, res: Response): void {
 	const token = jwtTokenFromRequest(req);
 
 	// ** Kill the session
