--- server/routes/v0/ping.ts
@@ -1,3 +1,0 @@
-import type { RequestHandler } from "express";
-
-export const ping: RequestHandler = (req, res) => res.json({ message: "Pong!" });
