+++ server/routes/v0/session/+server.ts
@@ -1,10 +1,10 @@
 import type { Request, Response } from "express";
-import { metadataFromRequest } from "../../auth/requireAuth.js";
-import { newAccessToken } from "../../auth/jwt.js";
-import { respondSuccess } from "../../responses.js";
-import { statsForUser } from "../../database/io.js";
+import { metadataFromRequest } from "../../../auth/requireAuth.js";
+import { newAccessToken } from "../../../auth/jwt.js";
+import { respondSuccess } from "../../../responses.js";
+import { statsForUser } from "../../../database/io.js";
 
-export async function getSession(req: Request, res: Response): Promise<void> {
+export async function GET(req: Request, res: Response): Promise<void> {
 	// ** If the user has the cookie set, respond with a JWT for the user
 
 	const metadata = await metadataFromRequest(req); // throws if bad
