+++ server/routes/v0/updateaccountid/+server.ts
@@ -1,19 +1,19 @@
 import type { Request, Response } from "express";
-import { BadRequestError, UnauthorizedError } from "../../errors/index.js";
-import { compare } from "../../auth/generators.js";
-import { generateTOTPSecretURI, verifyTOTP } from "../../auth/totp.js";
-import { respondSuccess } from "../../responses.js";
-import { upsertUser, userWithAccountId } from "../../database/io.js";
+import { BadRequestError, UnauthorizedError } from "../../../errors/index.js";
+import { compare } from "../../../auth/generators.js";
+import { generateTOTPSecretURI, verifyTOTP } from "../../../auth/totp.js";
+import { respondSuccess } from "../../../responses.js";
+import { upsertUser, userWithAccountId } from "../../../database/io.js";
 import { is, nonempty, optional, string, type } from "superstruct";
 
-const reqBody = type({
-	account: nonempty(string()),
-	newaccount: nonempty(string()),
-	password: nonempty(string()),
-	token: optional(nonempty(string())),
-});
+export async function POST(req: Request, res: Response): Promise<void> {
+	const reqBody = type({
+		account: nonempty(string()),
+		newaccount: nonempty(string()),
+		password: nonempty(string()),
+		token: optional(nonempty(string())),
+	});
 
-export async function postUpdateAccountId(req: Request, res: Response): Promise<void> {
 	if (!is(req.body, reqBody)) {
 		throw new BadRequestError("Improper parameter types");
 	}
