+++ server/routes/v0/version/+server.ts
@@ -1,5 +1,5 @@
 import type { RequestHandler } from "express";
-import { version } from "../../version.js";
+import { version } from "../../../version.js";
 
-export const serverVersion: RequestHandler = (req, res) =>
+export const GET: RequestHandler = (req, res) =>
 	res.json({ message: `Accountable v${version}`, version });
